% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagename.R
\name{packagename}
\alias{packagename}
\title{A function to state the name of the package}
\usage{
packagename()
}
\description{
A function that states the name of the package
}
\examples{
Type packagename() into the console to expand the acronym.
}
\keyword{Name}

